% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_results.R
\name{AddLogicGates}
\alias{AddLogicGates}
\title{extend model with node representing logic gate}
\usage{
AddLogicGates(child, logic, model)
}
\description{
extend model with node representing logic gate
}

