% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_results.R
\name{combiNEM.PlotResults}
\alias{combiNEM.PlotResults}
\title{Plots the winning pathway structure}
\usage{
combiNEM.PlotResults(results)
}
\arguments{
\item{LogiNEM}{object created by LogicNEM}
}
\description{
Plots the winning pathway structure
}

