% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LogiN.R
\name{combiNEM}
\alias{combiNEM}
\title{Logic NEMs - main function}
\usage{
combiNEM(filename = "random", method = "greedy", nIterations = 10,
  nModels = 0, random = list(single = 4, double = 1, reporters = 100, FPrate
  = 0.1, FNrate = 0.1, replicates = 1), plotsy = TRUE)
}
\arguments{
\item{filename}{A binary, tab-delimited matrix. Columns: single and double knockdowns. Rows: genes showing effect or not?
Default: random; artificial data is generated according to 'random' specifications}

\item{method}{greedy or exhaustive search. Default: greedy}

\item{nIterations}{number of iterations. Default: 10}

\item{nModels}{number of Models. Default: 0}

\item{random}{list specifying how the data should be generated: no. of single mutants, no. of double mutants, no. of reporterGenes,
FP-rate, FN-rate, no. of replicates}
}
\description{
Logic NEMs - main function
}
\author{
Madeline Diekmann
}
\seealso{
nem
}

