% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LogiN.R
\docType{data}
\name{combiNEMdata}
\alias{combiNEMdata}
\title{A binary effect matrix for a given set of knockouts.}
\format{A binary matrix with 2274 rows and 5 variables:
\describe{
  \item{row}{knockouts, E-Genes}
  \item{column}{S-Genes}
  ...
}}
\source{
\url{http://www.holstegelab.nl/publications/sv/signaling_redundancy/}
}
\usage{
combiNEMdata
}
\description{
A publicly availiable example data set for demonstrating the use of the combiNEM package.
For data preprocessing steps please check vignette.
}
\keyword{datasets}

