% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalFunctions.R
\name{CreateExtendedAdjacency}
\alias{CreateExtendedAdjacency}
\title{Create an extended adjacency matrix}
\usage{
CreateExtendedAdjacency(network, mutants, experiments)
}
\arguments{
\item{network}{network created by BoolNet from file}

\item{mutants}{vector of single knockouts}

\item{experiments}{vector of all knockouts}
}
\value{
extended adjacency matrix
}
\description{
extend adjacency matrices taking cycles and logics into account. For every given start state, the final state
is computed yu using BoolNet.
}
\examples{
library(BoolNet)
data(cellcycle)
extModel <- CreateExtendedAdjacency(cellcycle, c(cellcycle$genes, "CycD.Rb"), cellcycle$genes)
}

