% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalFunctions.R
\name{CreateRandomGraph}
\alias{CreateRandomGraph}
\title{Create a random graph}
\usage{
CreateRandomGraph(pathwayGenes, edgeProb = 0.5)
}
\arguments{
\item{pathwayGenes}{vector of genes in the pathway}

\item{edgeProb}{probability of random edge}
}
\value{
adjacency matrix
}
\description{
Returns a model graph with randomly sampled edges. Every possible edge has a probability to exist in the graph.
}
\examples{
graph <- CreateRandomGraph(c("Ikk1", "Ikk2", "RelA"))
}

