% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalFunctions.R
\name{CreateTopology}
\alias{CreateTopology}
\title{create topology for a randomly generated pathway topology}
\usage{
CreateTopology(single, double)
}
\arguments{
\item{single:}{number of single knockouts}

\item{double:}{number of double knockouts}
}
\description{
create topology for a randomly generated pathway topology
}

