% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_method.R
\name{EnumerateModels}
\alias{EnumerateModels}
\title{Enumerate all possible adjacency matrices}
\usage{
EnumerateModels(size, nodes = "none")
}
\arguments{
\item{size:}{number of desired nodes}
}
\description{
Returns a lst of all unique matrices
}

