% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\name{epi2bg}
\alias{epi2bg}
\title{Convert epiNEM model into general Boolean graph}
\usage{
epi2bg(t)
}
\arguments{
\item{t}{full epiNEM model}
}
\description{
Convert epiNEM model into general Boolean graph
}
\examples{
topology <- CreateTopology(3, 1, force = TRUE)
topology <- unlist(unique(topology), recursive = FALSE)
extTopology <- ExtendTopology(topology$model, 100)
b <- epi2bg(extTopology)
}
\author{
Madeline Diekmann
}
\seealso{
CreateTopology
}

