% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\name{epiAnno}
\alias{epiAnno}
\title{Plots logical gate data annotation.}
\usage{
epiAnno()
}
\value{
plot of heatmats showing the silencing scheme (expected data)
}
\description{
Plots logical gate data annotation.
}
\examples{
epiAnno()
}
\author{
Martin Pirkl
}
