% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\name{ExtendTopology}
\alias{ExtendTopology}
\title{Extending topology of normal "nem"}
\usage{
ExtendTopology(topology, nReporters)
}
\arguments{
\item{topology:}{model of a topology from CreateTopology}

\item{nReporters:}{number of effects reporters}
}
\description{
Extending topology of normal "nem"
}
\examples{
topology <- CreateTopology(3, 2, force = forcelogic)
topology <- unlist(unique(topology), recursive = FALSE)
extTopology <- ExtendTopology(topology$model, 100)
}
\author{
Madeline Diekmann
}
\seealso{
CreateTopology
}

