% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_method.R
\name{FindNeighbours}
\alias{FindNeighbours}
\title{find neighbors}
\usage{
FindNeighbours(model, edges)
}
\arguments{
\item{model:}{adjacency matrix of current model}

\item{edges:}{all non-self edges in the model}
}
\description{
Return all models that differ only one edge from the current model by removing or adding a non-self edge.
}

