% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalFunctions.R
\name{GenerateDoubleKO}
\alias{GenerateDoubleKO}
\title{generate a random double mutant for artificial data}
\usage{
GenerateDoubleKO(singleKOs)
}
\arguments{
\item{singleKOs:}{vector of single mutants}
}
\description{
generate a random double mutant for artificial data
}

