% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finalFunctions.R
\name{getExtendedAdjacency}
\alias{getExtendedAdjacency}
\title{create with logics extended adjacency matrix}
\usage{
getExtendedAdjacency(modelno, logicmatrix, column, adj, mutants, experiments)
}
\description{
create with logics extended adjacency matrix
}

