% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalFunctions.R
\name{getStarters}
\alias{getStarters}
\title{get starting vector}
\usage{
getStarters(mutants, experiments)
}
\arguments{
\item{mutants:}{vector of single knockouts}

\item{experiments:}{vector of all knockouts}
}
\description{
get mutant vectors for simulating boolean network
}

