% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_method.R
\name{GreedyHillClimber}
\alias{GreedyHillClimber}
\title{perform greedy hillclimbing algorithm}
\usage{
GreedyHillClimber(n, experiments, data2, data0, mutants)
}
\arguments{
\item{n:}{number of iterations}

\item{experiments:}{vector of mutants}

\item{data2:}{observed data}

\item{data0:}{complementary data2}
}
\description{
Performs iteration n of a greedy hillclimbing algorithm. Returns the score and the model representing
the optimal solution of the iteration.
}

