% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalFunctions.R
\name{includeLogic}
\alias{includeLogic}
\title{include logics into network. Returns an extended adjacency matrix.}
\usage{
includeLogic(adj, experiments, mutants)
}
\arguments{
\item{adj:}{adjacency matrix of graph}

\item{experiments:}{vector of all knockouts}

\item{mutants:}{vector of single knockouts}
}
\description{
include logics into network. Returns an extended adjacency matrix.
}

