% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finalFunctions.R
\name{MLL}
\alias{MLL}
\title{Evaluation of graphs}
\usage{
MLL(Phi, D1, D0)
}
\arguments{
\item{Phi:}{model to be evaluated}

\item{D1:}{observed data matrix}

\item{D0:}{complementary D1}
}
\description{
Computes marginal log-likelihood for model Phi given observed data matrix D1
}

