% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot.epiNEM}
\alias{plot.epiNEM}
\title{Plot pathway.}
\usage{
\method{plot}{epiNEM}(x, ...)
}
\arguments{
\item{x}{object of class epiNEM}

\item{...}{other arguments}
}
\value{
plot of the logical network
}
\description{
Plots the winning pathway structure
}
\examples{
data <- matrix(sample(c(0,1), 100*4, replace = TRUE), 100, 4)
colnames(data) <- c("A", "A.B", "B", "C")
rownames(data) <- paste("E", 1:100, sep = "_")
res <- epiNEM(data, method = "exhaustive")
plot(res)
}
