% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot.epiSim}
\alias{plot.epiSim}
\title{Plot simulations.}
\usage{
\method{plot}{epiSim}(x, ...)
}
\arguments{
\item{x}{object of class epiSim}

\item{...}{other arguments}
}
\value{
plot(s) of an epiNEM simulation analysis
}
\description{
Plots the simulation results
}
\examples{
res <- SimEpiNEM(runs = 1)
plot(res)
}
