% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\docType{data}
\name{sameith_GO}
\alias{sameith_GO}
\title{graph-based GO similarity scores, string GO annotations
for Sameith et al., 2015 data
The data consists of lists including epiNEM identified and
general similarity scores and GO annotations for each triple.
For details see the vignette.}
\description{
graph-based GO similarity scores, string GO annotations
for Sameith et al., 2015 data
The data consists of lists including epiNEM identified and
general similarity scores and GO annotations for each triple.
For details see the vignette.
}
\examples{
data(sameith_GO)
}
