% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\docType{data}
\name{sameith_res}
\alias{sameith_res}
\title{Example data: epiNEM results for the Sameith et al., 2015 knock-out screen}
\description{
Example data: epiNEM results for the Sameith et al., 2015 knock-out screen
}
\examples{
data(sameith_res)
}

