% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\docType{data}
\name{sameith_string}
\alias{sameith_string}
\title{Example data: string db interactions of interactions identified by epiNEM for the Sameith et al., 2010 knock-out screen}
\description{
Example data: string db interactions of interactions identified by epiNEM for the Sameith et al., 2010 knock-out screen
}
\examples{
data(sameith_string)
}

