% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\docType{data}
\name{wageningen_GO}
\alias{wageningen_GO}
\title{graph-based GO similarity scores, string GO annotations
for van Wageningen et al., 2015 data}
\description{
The data consists of lists including epiNEM identified and
general similarity scores and GO annotations for each triple.
For details see the vignette.
}
\examples{
data(wageningen_GO)
}
