% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\docType{data}
\name{wageningen_res}
\alias{wageningen_res}
\title{Example data: epiNEM results for the Wageningen et al., 2010 knock-out screen}
\description{
Example data: epiNEM results for the Wageningen et al., 2010 knock-out screen
}
\examples{
data(wageningen_res)
}

