% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{AddLogicGates}
\alias{AddLogicGates}
\title{Add logic.}
\usage{
AddLogicGates(child, logic, model)
}
\arguments{
\item{child}{define the child}

\item{logic}{define the logical gate}

\item{model}{normal model}
}
\value{
model list with additional logic gate
}
\description{
extend model with node representing logic gate
}
\examples{
model <- CreateRandomGraph(c("Ikk1", "Ikk2", "RelA"))
model2 <- AddLogicGates("RelA", "OR", model)
}
