% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/finalFunctions.R
\name{createExtendedAdjacency}
\alias{createExtendedAdjacency}
\title{Create an extended adjacency matrix}
\usage{
createExtendedAdjacency(network, mutants, experiments)
}
\arguments{
\item{network:}{network created by BoolNet from file}

\item{mutants:}{vector of single knockouts}

\item{experiments:}{vector of all knockouts}
}
\description{
extend adjacency matrices taking cycles and logics into account. For every given start state, the final state
is computed yu using BoolNet.
}

