% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{epiNEM.PlotResults}
\alias{epiNEM.PlotResults}
\title{Plots the winning pathway structure}
\usage{
epiNEM.PlotResults(results)
}
\arguments{
\item{results}{object created by epiNEM}
}
\description{
Plots the winning pathway structure
}
\examples{
res <- epiNEM()
epiNEM.PlotResults(res)
}

