% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\name{epiNEM}
\alias{epiNEM}
\title{Epistatic NEMs - main function}
\usage{
epiNEM(filename = "random", method = "greedy", nIterations = 10,
  nModels = 0, random = list(single = 4, double = 1, reporters = 100, FPrate
  = 0.1, FNrate = 0.1, replicates = 1), plotsy = TRUE, ltype = "marginal",
  para = c(0.13, 0.05))
}
\arguments{
\item{filename}{A binary, tab-delimited matrix. Columns: single and double knockdowns. Rows: genes showing effect or not?
Default: random; artificial data is generated according to 'random' specifications}

\item{method}{greedy or exhaustive search. Default: greedy}

\item{nIterations}{number of iterations. Default: 10}

\item{nModels}{number of Models. Default: 0}

\item{random}{list specifying how the data should be generated: no. of single mutants, no. of double mutants, no. of reporterGenes,
FP-rate, FN-rate, no. of replicates}

\item{plotsy}{atm not used}

\item{ltype}{likelihood either "marginal" or "maximum"}

\item{para}{false positive and false negative rates}
}
\description{
Epistatic NEMs - main function
}
\examples{
res <- epiNEM()
}
\author{
Madeline Diekmann
}
\seealso{
nem
}

