% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\name{EpiNEM2BooleanGraph}
\alias{EpiNEM2BooleanGraph}
\title{Convert epiNEM model into general Boolean graph.
Only needed for comparing accuracy of inferred network for bnem and epiNEM.
Needs the package bnem from https://github.com/MartinFXP/bnem.
In the future bnem will be submitted to bioconductor.}
\usage{
EpiNEM2BooleanGraph(t)
}
\arguments{
\item{t}{full epiNEM model}
}
\value{
boolean hyper-graph
}
\description{
Convert epiNEM model into general Boolean graph.
Only needed for comparing accuracy of inferred network for bnem and epiNEM.
Needs the package bnem from https://github.com/MartinFXP/bnem.
In the future bnem will be submitted to bioconductor.
}
\examples{
topology <- CreateTopology(3, 1, force = TRUE)
topology <- unlist(unique(topology), recursive = FALSE)
extTopology <- ExtendTopology(topology$model, 100)
b <- EpiNEM2BooleanGraph(extTopology)
}
\author{
Madeline Diekmann
}
\seealso{
CreateTopology
}

