% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\name{GenerateData}
\alias{GenerateData}
\title{Generate data from extended model}
\usage{
GenerateData(model, extTopology, FPrate, FNrate, replicates)
}
\arguments{
\item{model}{model of a topology from CreateTopology}

\item{extTopology}{extended topology}

\item{FPrate}{false positive rate}

\item{FNrate}{false negative rate}

\item{replicates}{number of replicates}
}
\description{
Generate data from extended model
}
\examples{
topology <- CreateTopology(3, 2, force = forcelogic)
topology <- unlist(unique(topology), recursive = FALSE)
extTopology <- ExtendTopology(topology$model, 100)
sortedData <- GenerateData(topology$model, extTopology, 0.05, 0.13, 3)
}
\author{
Madeline Diekmann
}
\seealso{
CreateTopology
}

