% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot.epiNEM}
\alias{plot.epiNEM}
\title{Plots the winning pathway structure}
\usage{
\method{plot}{epiNEM}(x, ...)
}
\arguments{
\item{results}{object of class epiNEM}
}
\value{
plot of the logical network
}
\description{
Plots the winning pathway structure
}
\examples{
res <- epiNEM()
plot(res)
}

