% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\docType{data}
\name{sameith_string}
\alias{sameith_string}
\title{sig. of string interaction scores for Sameith et al., 2015 data}
\description{
sig. of string interaction scores for Sameith et al., 2015 data
}
\examples{
data(sameith_string)
}

