% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\docType{data}
\name{wageningen_string}
\alias{wageningen_string}
\title{sig. of string interaction scores
for van Wageningen et al., 2010 data}
\description{
sig. of string interaction scores
for van Wageningen et al., 2010 data
}
\examples{
data(wageningen_string)
}

