% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiN.R
\docType{data}
\name{wageningen_string}
\alias{wageningen_string}
\title{sig. of string interaction scores
for van Wageningen et al., 2010 data
The data consists of a list including a vectors of pairs (for interactions)
and a corresponding list of interaction scores derived form the
string database.
For details see the vignette.}
\description{
sig. of string interaction scores
for van Wageningen et al., 2010 data
The data consists of a list including a vectors of pairs (for interactions)
and a corresponding list of interaction scores derived form the
string database.
For details see the vignette.
}
\examples{
data(wageningen_string)
}

